AXMARIL/requirements.txt
altgraph==0.17.4
aniso8601==10.0.1
anyio==4.0.0
asttokens==3.0.0
attrs==23.2.0
Authlib==1.2.0
autopep8==2.3.2
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.13.5
bleach==6.2.0
cachelib==0.13.0
cachetools==5.3.1
captcha==0.6.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cryptography==39.0.2
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
enum-compat==0.0.3
exceptiongroup==1.1.3
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.13.1
flasgger==0.9.7b2
Flask==2.2.2
Flask-Caching==2.3.1
Flask-Cors==3.0.10
flask-marshmallow==1.2.0
flask-restx==1.3.0
flask-sock==0.7.0
flask-swagger-ui==4.11.1
gevent==24.2.1
google-api-core==2.12.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.61.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
ldap3==2.9.1
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib-inline==0.1.7
mistune==3.0.2
mongita==1.2.0
mysql-connector-python==9.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
oauthlib==3.2.2
packaging==23.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.5
password-strength==0.0.3.post2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.4.0
prompt_toolkit==3.0.51
protobuf==4.24.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.14.0
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.19.2
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.4.0
PyKMIP==0.10.0
pymongo==4.14.1
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyotp==2.7.0
pyparsing==3.1.1
pysyncobj==0.3.14
pytest==8.0.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2025.2
PyYAML==6.0.1
pyzmq==27.0.2
referencing==0.33.0
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.18.0
rsa==4.9
shortuuid==1.0.11
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
tempfile2==0.1.2
tinycss2==1.4.0
tldextract==5.1.1
tomli==2.0.1
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==1.26.18
urlquote==2.0.0
validate-email==1.3
validators==0.22.0
waitress==2.1.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
yarg==0.1.9
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
